Refactor HeadingElement Component logic: improved event managing. Now we don't need the cleanup function to update the recent selected element's data. Instead we rely on 'Event Propagation rules'.
We are using Event capturing for the 'Click Outside' handler on the HeadingElement and default propagation which is Event bubbling for the withConfig hoc's parent div. So that click outside handler invoke first and withConfig select element invoke second.

This improves the amount of updates. With the previous logic 1 extra unnecessary updates was firing. Now updates are perfect.